1. V4L2 capture now doesn't depend on Qt 2. Unpmerging some closed code 3. Removing substandard unfixable code

+++ core/fileformats/objLoader.cpp
@@ -316,7 +316,7 @@ int OBJLoader::loadOBJSimple(istream &input, Mesh3D &mesh)
 
             vector<int32_t> face;
 
-            for (int i = 0; i < strs.size(); i++)
+            for (size_t i = 0; i < strs.size(); i++)
             {
                 // LOCAL_PRINT(("Attribute: %s\n", strs[i].c_str()));
                 std::stringstream splitter(strs[i]);
